requirements.txt
# 深度学习框架
torch>=2.0.0
transformers>=4.35.0
accelerate>=0.24.0
bitsandbytes>=0.41.0

# 检索相关
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
whoosh>=2.7.4

# Reranker
FlagEmbedding>=1.2.0

# 工具库
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0
pydantic>=2.0.0
pyyaml>=6.0.0

# 缓存（可选）
redis>=5.0.0

# 监控和日志
loguru>=0.7.0
psutil>=5.9.0

# API服务
fastapi>=0.104.0
uvicorn>=0.24.0
